- unpack object size as unsigned int32 with the hope we can deserialise things   larger than 2GB then - patch by Christian Bühler
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@16492 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libgwyddion/gwyserializable.c
@@ -1361,7 +1361,7 @@ gwy_deserialize_object_hash(const guchar *buffer,
     g_return_val_if_fail(mysize, NULL);
     *position += mysize;
 
-    mysize = gwy_deserialize_int32(buffer, size, position);
+    mysize = (guint32)gwy_deserialize_int32(buffer, size, position);
     g_return_val_if_fail(mysize <= size - *position, NULL);
     items = gwy_deserialize_hash_items(buffer + *position, mysize, nitems);
     *position += mysize;
